+++ R/LearnerClassifKeras.R
@@ -21,7 +21,7 @@
 #' Some exceptions are documented here.
 #' * `model`: A compiled keras model.
 #' * `class_weight`: needs to be a named list of class-weights
-#'   for the dierent classes numbered from 0 to c-1 (for c classes).
+#'   for the different classes numbered from 0 to c-1 (for c classes).
 #'   ```
 #'   Example:
 #'   wts = c(0.5, 1)
@@ -51,7 +51,15 @@
 LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassif,
   public = list(
     architecture = NULL,
-    initialize = function(architecture = KerasArchitectureCustomModel$new()) {
+    initialize = function(
+        id = "classif.keras",
+        predict_types = c("response", "prob"),
+        feature_types = c("integer", "numeric"),
+        properties = c("twoclass", "multiclass"),
+        packages = "keras",
+        man = "mlr3keras::mlr_learners_classif.keras",
+        architecture = KerasArchitectureCustomModel$new()
+      ) {
       self$architecture = assert_class(architecture, "KerasArchitecture")
       ps = ParamSet$new(list(
         ParamInt$new("epochs", default = 30L, lower = 1L, tags = "train"),
@@ -64,20 +72,19 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
         ParamInt$new("verbose", lower = 0L, upper = 1L, tags = c("train", "predict"))
       ))
       ps$values = list(epochs = 30L, callbacks = list(), validation_split = 1/3, batch_size = 128L, low_memory = FALSE)
-      ps = ParamSetCollection$new(list(ps, self$architecture$param_set))
+
       super$initialize(
-        id = "classif.keras",
-        param_set = ps,
-        predict_types = c("response", "prob"),
-        feature_types = c("integer", "numeric"),
-        properties = c("twoclass", "multiclass"),
-        packages = "keras",
-        man = "mlr3keras::mlr_learners_classif.keras"
+        id = assert_character(id, len = 1),
+        param_set = ParamSetCollection$new(list(ps, self$architecture$param_set)),
+        predict_types = assert_character(predict_types),
+        feature_types = assert_character(feature_types),
+        properties = assert_character(properties),
+        packages = assert_character(packages),
+        man = assert_character(man)
       )
 
       # Set y_transform
-      self$architecture$set_transform(
-        "y",
+      self$architecture$set_transform("y",
         function(target, pars, model_loss) {
           y = to_categorical(as.integer(target) - 1)
           if (model_loss == "binary_crossentropy") y = y[, 1, drop = FALSE]
@@ -110,12 +117,13 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
           verbose = pars$verbose,
           callbacks = pars$callbacks)
 
+
       } else {
         # Validation split
         rho = rsmp("holdout", ratio = 1 - pars$validation_split)
         rho$instantiate(task)
 
-        train_gen <- make_data_generator(
+        train_gen = make_data_generator(
           task = task,
           batch_size = pars$batch_size,
           filter_ids = rho$train_set(1),
@@ -123,7 +131,7 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
           y_transform = function(y) {self$architecture$transforms$y(y, pars, model$loss)}
         )
 
-        valid_gen <- make_data_generator(
+        valid_gen = make_data_generator(
           task = task,
           batch_size = pars$batch_size,
           filter_ids = rho$test_set(1),
@@ -132,8 +140,8 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
         )
 
         # Number of steps
-        train_steps <- ceiling(length(rho$train_set(1)) / pars$batch_size)
-        valid_steps <- ceiling(length(rho$test_set(1)) / pars$batch_size)
+        train_steps = ceiling(length(rho$train_set(1)) / pars$batch_size)
+        valid_steps = ceiling(length(rho$test_set(1)) / pars$batch_size)
 
         # Train with generator
         if(pars$validation_split > 0) {
@@ -166,29 +174,36 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
 
       features = task$data(cols = task$feature_names)
       newdata = self$architecture$transforms$x(features, pars)
+      pars = pars[intersect(names(pars), self$keras_predict_pars)]
+      response = prob = NULL
 
       if (inherits(self$model$model, "keras.engine.sequential.Sequential")) {
         if (self$predict_type == "response") {
-          p = invoke(keras::predict_classes, self$model$model, x = newdata, .args = pars)
-          p = factor(self$model$class_names[p + 1])
-          PredictionClassif$new(task = task, response = drop(p))
+          response = invoke(keras::predict_classes, self$model$model, x = newdata, .args = pars)
+          response = drop(factor(self$model$class_names[response + 1]))
         } else if (self$predict_type == "prob") {
-          prob = invoke(keras::predict_proba, self$model$model, x = newdata, .args = pars)
-          if (ncol(prob) == 1L) prob = cbind(1-prob, prob)
-          colnames(prob) = task$class_names
-          PredictionClassif$new(task = task, prob = prob)
+          p = invoke(keras::predict_proba, self$model$model, x = newdata, .args = pars)
+          if (ncol(p) == 1L) {
+            if (task$class_names[1] != task$positive) p = cbind(1 - p, p)
+            else p = cbind(p, 1 - p)
+          }
         }
       } else {
         p = invoke(self$model$model$predict, x = newdata, .args = pars)
+          if (ncol(p) == 1L) {
+            if (task$class_names[1] != task$positive) p = cbind(1 - p, p)
+            else p = cbind(p, 1 - p)
+          }
         if (self$predict_type == "response") {
-          p = factor(self$model$class_names[apply(p, 1, which.max)])
-          PredictionClassif$new(task = task, response = drop(p))
-        } else if (self$predict_type == "prob") {
-          if (ncol(p) == 1L) p = cbind(1-p, p)
-          colnames(p) = task$class_names
-          PredictionClassif$new(task = task, prob = p)
+          response = factor(self$model$class_names[apply(p, 1, which.max)])
         }
       }
+      if (self$predict_type == "prob") {
+        prob = p
+        colnames(prob) = task$class_names
+      }
+      PredictionClassif$new(task = task, prob = prob, response = response)
+
     },
     save = function(filepath) {
       assert_path_for_output(filepath)
@@ -198,6 +213,7 @@ LearnerClassifKeras = R6::R6Class("LearnerClassifKeras", inherit = LearnerClassi
     plot = function() {
       if (is.null(self$model)) stop("Model must be trained before saving")
       plot(self$model$history)
-    }
+    },
+    keras_predict_pars = c("batch_size", "verbose")
   )
 )