+++ R/LearnerKerasFF.R
@@ -21,6 +21,9 @@ LearnerClassifKerasFF = R6::R6Class("LearnerClassifKerasFF",
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamLgl$new("use_embedding", default = TRUE, tags = c("train", "predict")),
+        ParamDbl$new("embed_dropout", default = 0, lower = 0, upper = 1, tags = "train"),
+        ParamDbl$new("embed_size", default = NULL, lower = 1, upper = Inf, tags = "train", special_vals = list(NULL)),
         ParamUty$new("layer_units", default = c(32, 32, 32), tags = "train"),
         ParamUty$new("initializer", default = "initializer_glorot_uniform()", tags = "train"),
         ParamUty$new("regularizer", default = "regularizer_l1_l2()", tags = "train"),
@@ -37,6 +40,7 @@ LearnerClassifKerasFF = R6::R6Class("LearnerClassifKerasFF",
         ParamUty$new("metrics", tags = "train")
       ))
       ps$values = list(
+        use_embedding = TRUE, embed_dropout = 0, embed_size = NULL,
         activation = "relu",
         layer_units = c(32, 32, 32),
         initializer = initializer_glorot_uniform(),
@@ -48,8 +52,11 @@ LearnerClassifKerasFF = R6::R6Class("LearnerClassifKerasFF",
         metrics = "accuracy",
         output_activation = "softmax"
       )
-      arch = KerasArchitectureFF$new(build_arch_fn = build_keras_ff_model,  param_set = ps)
-      super$initialize(architecture = arch)
+      arch = KerasArchitectureFF$new(build_arch_fn = build_keras_ff_model, param_set = ps)
+      super$initialize(
+        feature_types = c("integer", "numeric", "factor", "ordered"),
+        architecture = arch
+      )
     }
   )
 )
@@ -76,6 +83,9 @@ LearnerRegrKerasFF = R6::R6Class("LearnerRegrKerasFF",
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamLgl$new("use_embedding", default = TRUE, tags = c("train", "predict")),
+        ParamDbl$new("embed_dropout", default = 0, lower = 0, upper = 1, tags = "train"),
+        ParamDbl$new("embed_size", default = NULL, lower = 1, upper = Inf, tags = "train", special_vals = list(NULL)),
         ParamUty$new("layer_units", default = c(32, 32, 32), tags = "train"),
         ParamUty$new("initializer", default = "initializer_glorot_uniform()", tags = "train"),
         ParamUty$new("regularizer", default = "regularizer_l1_l2()", tags = "train"),
@@ -93,6 +103,7 @@ LearnerRegrKerasFF = R6::R6Class("LearnerRegrKerasFF",
         ParamUty$new("metrics", default = "mean_squared_logarithmic_error", tags = "train")
       ))
       ps$values = list(
+        use_embedding = TRUE, embed_dropout = 0,  embed_size = NULL,
         activation = "relu",
         layer_units = c(32, 32, 32),
         initializer = initializer_glorot_uniform(),
@@ -104,11 +115,33 @@ LearnerRegrKerasFF = R6::R6Class("LearnerRegrKerasFF",
         metrics = "mean_squared_logarithmic_error",
         output_activation = "linear"
       )
-      arch = KerasArchitectureFF$new(build_arch_fn = build_keras_ff_model,  param_set = ps)
-      super$initialize(architecture = arch)
+      arch = KerasArchitectureFF$new(build_arch_fn = build_keras_ff_model, param_set = ps)
+      super$initialize(
+        feature_types = c("integer", "numeric", "factor", "ordered"),
+        architecture = arch
+      )
     }
   )
 )
+
+#' @title Keras Neural Network Feed Forward architecture
+#' @rdname KerasArchitecture
+#' @family KerasArchitectures
+#' @export
+KerasArchitectureFF = R6::R6Class("KerasArchitectureFF",
+  inherit = KerasArchitecture,
+  public = list(
+    initialize = function(build_arch_fn, param_set) {
+      super$initialize(build_arch_fn = build_arch_fn, param_set = param_set,
+        x_transform = function(features, pars) {
+          if (pars$use_embedding) keras_array(reshape_data_embedding(features)$data)
+          else as.matrix(model.matrix(~. - 1 , features))
+      })
+    }
+  )
+)
+
+
 
 # Builds a Keras Feed Forward Neural Network
 # @param task [`Task`] \cr
@@ -131,8 +164,15 @@ build_keras_ff_model = function(task, pars) {
     }
   }
 
-  model = keras_model_sequential()
-  if (pars$use_dropout) model = model %>% layer_dropout(pars$input_dropout, input_shape = input_shape)
+  if (pars$use_embedding) {
+    embd = make_embedding(task, pars$embed_size, pars$embed_dropout)
+    model = embd$layers
+  } else {
+    model = keras_model_sequential()
+  }
+
+  if (pars$use_dropout)
+    model = model %>% layer_dropout(pars$input_dropout, input_shape = input_shape)
 
   # Build hidden layers
   for (i in seq_len(length(pars$layer_units))) {
@@ -156,6 +196,8 @@ build_keras_ff_model = function(task, pars) {
     model = model %>% layer_dense(units = output_shape, activation = "sigmoid")
   else
     model = model %>% layer_dense(units = output_shape, activation = pars$output_activation)
+
+  if (pars$use_embedding) model = keras_model(inputs = embd$inputs, outputs = model)
 
   model %>% compile(
     optimizer = pars$optimizer,