+++ R/LearnerTabNet.R
@@ -25,6 +25,7 @@ LearnerClassifTabNet = R6::R6Class("LearnerClassifTabNet",
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamInt$new("embed_size", default = NULL, lower = 1L, upper = Inf, tags = "train", special_vals = list(NULL)),
         ParamLgl$new("stacked", default = FALSE, tags = "train"),
         ParamInt$new("num_layers", lower = 1, upper = Inf, default = 1L, tags = "train"),
         ParamDbl$new("batch_momentum", lower = 0, upper = 1, tags = "train"),
@@ -44,6 +45,7 @@ LearnerClassifTabNet = R6::R6Class("LearnerClassifTabNet",
       ))
       ps$add_dep("num_layers", "stacked", CondEqual$new(TRUE))
       ps$values = list(
+        embed_size = NULL,
         stacked = FALSE,
         batch_momentum = 0.98,
         relaxation_factor = 1.0,
@@ -59,9 +61,13 @@ LearnerClassifTabNet = R6::R6Class("LearnerClassifTabNet",
         loss = "categorical_crossentropy",
         metrics = "accuracy"
       )
-
       arch = KerasArchitectureTabNet$new(build_arch_fn = build_keras_tabnet, param_set = ps)
-      super$initialize(architecture = arch)
+      super$initialize(
+        id = "classif.tabnet",
+        feature_types = c("integer", "numeric", "factor", "logical"),
+        packages = c("keras", "tensorflow", "reticulate"),
+        man = "mlr3keras::mlr_learners_classif.tabnet",
+        architecture = arch)
       self$param_set$values$validation_split = 0 # Does not to work with tf_data.
     }
   )
@@ -95,6 +101,7 @@ LearnerRegrTabNet = R6::R6Class("LearnerRegrTabNet",
   public = list(
     initialize = function() {
       ps = ParamSet$new(list(
+        ParamInt$new("embed_size", default = NULL, lower = 1L, upper = Inf, tags = "train", special_vals = list(NULL)),
         ParamLgl$new("stacked", default = FALSE, tags = "train"),
         ParamInt$new("num_layers", lower = 1, upper = Inf, default = 1L, tags = "train"),
         ParamDbl$new("batch_momentum", lower = 0, upper = 1, tags = "train"),
@@ -115,6 +122,7 @@ LearnerRegrTabNet = R6::R6Class("LearnerRegrTabNet",
       ))
       ps$add_dep("num_layers", "stacked", CondEqual$new(TRUE))
       ps$values = list(
+        embed_size = NULL,
         stacked = FALSE,
         batch_momentum = 0.98,
         relaxation_factor = 1.0,
@@ -132,7 +140,12 @@ LearnerRegrTabNet = R6::R6Class("LearnerRegrTabNet",
       )
 
       arch = KerasArchitectureTabNet$new(build_arch_fn = build_keras_tabnet, param_set = ps)
-      super$initialize(architecture = arch)
+      super$initialize(
+        id = "regr.tabnet",
+        feature_types = c("integer", "numeric", "factor", "logical"),
+        packages = c("keras", "tensorflow", "reticulate"),
+        man = "mlr3keras::mlr_learners_regr.tabnet",
+        architecture = arch)
       self$param_set$values$validation_split = 0 # Does not to work with tf_data.
     }
   )
@@ -148,7 +161,15 @@ KerasArchitectureTabNet = R6::R6Class("KerasArchitectureTabNet",
   public = list(
     initialize = function(build_arch_fn, x_transform, y_transform, param_set) {
       x_transform = function(features, pars) {
-        x = lapply(names(features), function(x) {as.matrix(features[, get(x)])})
+        x = lapply(names(features), function(x) {
+          x = features[, get(x)]
+          if (is.numeric(x) || is.integer(x)) {
+            as.matrix(as.numeric(x))
+          } else if (is.logical(x)) {
+            as.matrix(as.integer(x))
+          } else {
+            as.matrix(x)
+          }})
         names(x) = names(features)
         return(x)
       }
@@ -166,8 +187,7 @@ build_keras_tabnet = function(task, pars) {
       keras::install_keras(extra_packages = c('tensorflow-hub', 'tabnet==0.1.4.1')).")
   }
   tabnet = reticulate::import("tabnet")
-
-  feature_columns = make_tf_feature_cols(task)
+  feature_columns = make_tf_feature_cols(task, pars$embed_size)
   tabnet_param_names = c("feature_dim", "output_dim", "num_decision_steps", "relaxation_factor",
     "sparsity_coefficient", "virtual_batch_size", "norm_type", "num_groups")
   if (pars$stacked) tabnet_param_names = c(tabnet_param_names, "num_layers")
@@ -176,12 +196,14 @@ build_keras_tabnet = function(task, pars) {
     if (pars$stacked) clf = tabnet$StackedTabNetClassifier
     else clf = tabnet$TabNetClassifier
     model = invoke(clf,
+      num_features = get_tf_num_features(task, pars),
       feature_columns = feature_columns, num_classes = length(task$class_names),
       .args = pars[tabnet_param_names])
   } else if (inherits(task, "TaskRegr")) {
     if (pars$stacked) regr = tabnet$StackedTabNetRegressor
     else regr = tabnet$TabNetRegressor
     model = invoke(regr,
+      num_features = get_tf_num_features(task, pars),
       feature_columns = feature_columns, num_regressors = 1L,
       .args = pars[tabnet_param_names])
   }
@@ -193,13 +215,56 @@ build_keras_tabnet = function(task, pars) {
   )
 }
 
+# Create a tf$feature_column according to the type of a Task's column.
 # FIXME: This covers only the most basic feature types, needs to be extended.
-make_tf_feature_cols = function(task) {
-  assert_class(task, "Task")
-  make_feature_column = function(id, type) {
-    if (type == "numeric") tensorflow::tf$feature_column$numeric_column(id)
-    else if (type == "integer") tensorflow::tf$feature_column$numeric_column(id)
-    else tensorflow::tf$feature_column$categorical_column_with_vocabulary_list(id, task$levels(id)[[1]])
+make_tf_feature_column = function(id, type, args) {
+  if (type %in% c("numeric", "integer")) tensorflow::tf$feature_column$numeric_column(id)
+  else if (type == "logical") {
+    tensorflow::tf$feature_column$indicator_column(
+      tensorflow::tf$feature_column$categorical_column_with_identity(id, num_buckets = 2L)
+    )
   }
-  feature_columns = pmap(.f = make_feature_column, .x = task$feature_types)
+  else if (type %in% c("factor", "character", "ordered")) {
+    if (is.null(args$embed_size)) args$embed_size = get_default_embed_size(args$levels[[id]])
+    tensorflow::tf$feature_column$embedding_column(
+      tensorflow::tf$feature_column$categorical_column_with_vocabulary_list(id, args$levels[[id]]),
+      dimension = as.integer(args$embed_size)
+    )
+  }
 }
+
+get_tf_num_features = function(task, pars) {
+  dims = pmap_int(task$feature_types, function(id, type, levels) {
+    switch(type,
+      "numeric" = 1L,
+      "integer" = 1L,
+      "logical" = 2L,
+      "factor" = {
+        if (is.null(pars$embed_size))
+          get_default_embed_size(levels[[1]][[id]])
+        else pars$embed_size
+      },
+      "character" = {
+        if (is.null(pars$embed_size))
+          get_default_embed_size(levels[[1]][[id]])
+        else pars$embed_size
+      },
+      "ordered" = {
+        if (is.null(pars$embed_size))
+          get_default_embed_size(levels[[1]][[id]])
+        else pars$embed_size
+      }
+  )}, list(task$levels()))
+  sum(dims)
+}
+
+make_tf_feature_cols = function(task, embed_size = NULL) {
+  assert_r6(task, "Task")
+  feature_columns = pmap(.f = make_tf_feature_column, .x = task$feature_types, list(levels = task$levels(), embed_size = embed_size))
+}
+
+get_default_embed_size = function(levels) {
+    # As a default we use the fast.ai heuristic
+    as.integer(min(600L, round(1.6 * length(levels)^0.56)))
+}
+