+++ NEWS.md
@@ -1,7 +1,13 @@
+# mlr3keras 0.1.2
+
+## Learners
+* TabNet and FeedForward can now deal with factor / ordered / character features
+* FeedForward Keras Models now default to "embeddings" for factor features
+
 # mlr3keras 0.1.1
 
 ## General
-* KerasArchitecture: 
+* KerasArchitecture:
   Introduced new abstraction for architectures.
   This should rarely be visible to users but makes stuff easier to extend.